De-dubplicate function that performs biorthogonal transformation.
