+++ examples/stcc/stcc.cpp
@@ -8,14 +8,8 @@
 
 #include <clocale>
 
-#include <Eigen/Eigenvalues>
-
 using namespace sequant;
 
-ExprPtr biorthogonal_transform(
-    const sequant::ExprPtr& expr, int n_particles,
-    const std::vector<std::vector<sequant::Index>>& ext_index_groups = {{}},
-    double threshold = 1.e-12);
 ExprPtr symmetrize_expr(
     ExprPtr& expr,
     const container::vector<container::vector<Index>>& ext_index_groups = {{}});
@@ -125,117 +119,6 @@ int main(int argc, char* argv[]) {
         runtime_assert(cc_st_r.at(2)->size() == 73)   // T2
         runtime_assert(cc_st_r.at(3)->size() == 490)  // T3
   }
-}
-
-ExprPtr biorthogonal_transform(
-    const sequant::ExprPtr& expr, const int n_particles,
-    const std::vector<std::vector<sequant::Index>>& ext_index_groups,
-    const double threshold) {
-  assert(n_particles != 0);
-  assert(!ext_index_groups.empty());
-
-  using sequant::container::svector;
-
-  // Coefficients
-  std::vector<double> bt_coeff_vec;
-  {
-    using namespace Eigen;
-    // Dimension of permutation matrix is n_particles!
-    const auto n = sequant::factorial(n_particles);
-
-    // Permutation matrix
-    Eigen::Matrix<double, Dynamic, Dynamic> M(n, n);
-    {
-      M.setZero();
-      size_t n_row = 0;
-      svector<int, 6> v(n_particles), v1(n_particles);
-      std::iota(v.begin(), v.end(), 0);
-      std::iota(v1.begin(), v1.end(), 0);
-      do {
-        std::vector<double> permutation_vector;
-        do {
-          permutation_vector.push_back(
-              std::pow(-2, sequant::count_cycles(v1, v)));
-        } while (std::next_permutation(v.begin(), v.end()));
-        Eigen::VectorXd pv_eig = Eigen::Map<Eigen::VectorXd, Eigen::Unaligned>(
-            permutation_vector.data(), permutation_vector.size());
-        M.row(n_row) = pv_eig;
-        ++n_row;
-      } while (std::next_permutation(v1.begin(), v1.end()));
-      M *= std::pow(-1, n_particles);
-    }
-
-    // Normalization constant
-    double scalar;
-    {
-      auto nonZero = [&threshold](const double& d) {
-        return abs(d) > threshold;
-      };
-
-      // Solve system of equations
-      SelfAdjointEigenSolver<MatrixXd> eig_solver(M);
-      std::vector<double> eig_vals(eig_solver.eigenvalues().size());
-      VectorXd::Map(&eig_vals[0], eig_solver.eigenvalues().size()) =
-          eig_solver.eigenvalues();
-
-      double non0count =
-          std::count_if(eig_vals.begin(), eig_vals.end(), nonZero);
-      scalar = eig_vals.size() / non0count;
-    }
-
-    // Find Pseudo Inverse, get 1st row only
-    MatrixXd pinv = M.completeOrthogonalDecomposition().pseudoInverse();
-    bt_coeff_vec.resize(pinv.rows());
-    VectorXd::Map(&bt_coeff_vec[0], bt_coeff_vec.size()) = pinv.row(0) * scalar;
-  }
-
-  // Transformation maps
-  std::vector<std::map<Index, Index>> bt_maps;
-  {
-    std::vector<Index> idx_list(ext_index_groups.size());
-
-    for (auto i = 0; i != ext_index_groups.size(); ++i) {
-      idx_list[i] = *ext_index_groups[i].begin();
-    }
-
-    const std::vector<Index> const_idx_list = idx_list;
-
-    do {
-      std::map<Index, Index> map;
-      auto const_list_ptr = const_idx_list.begin();
-      for (auto& i : idx_list) {
-        map.emplace(std::make_pair(*const_list_ptr, i));
-        const_list_ptr++;
-      }
-      bt_maps.push_back(map);
-    } while (std::next_permutation(idx_list.begin(), idx_list.end()));
-  }
-
-  // If this assertion fails, change the threshold parameter
-  assert(bt_coeff_vec.size() == bt_maps.size());
-
-  // Checks if the replacement map is a canonical sequence
-  auto is_canonical = [](const std::map<Index, Index>& idx_map) {
-    bool canonical = true;
-    for (auto&& pair : idx_map)
-      if (pair.first != pair.second) return false;
-    return canonical;
-  };
-
-  // Scale transformed expressions and append
-  Sum bt_expr{};
-  auto coeff_it = bt_coeff_vec.begin();
-  for (auto&& map : bt_maps) {
-    const auto v = to_rational(*coeff_it, threshold);
-    if (is_canonical(map))
-      bt_expr.append(ex<Constant>(v) * expr->clone());
-    else
-      bt_expr.append(ex<Constant>(v) *
-                     sequant::transform_expr(expr->clone(), map));
-    coeff_it++;
-  }
-  ExprPtr result = std::make_shared<Sum>(bt_expr);
-  return result;
 }
 
 // Generate S operator from external index list
